backend/requirements.txt
# CareConnect Backend Requirements - Updated for Qloo integration

# Core FastAPI stack - compatible versions
fastapi==0.116.1 # Previously updated
uvicorn==0.34.0
pydantic==2.11.1

# Environment and configuration
python-dotenv==1.0.0

# HTTP client libraries
httpx>=0.28.1
requests==2.32.4
aiohttp==3.9.0

# Google APIs
google-cloud-vision==3.4.5
google-adk==1.7.0

# Logging and utilities
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS support
python-multipart>=0.0.9

# Date/time utilities
python-dateutil==2.9.0.post0 # <--- UPDATED THIS LINE

# JSON handling
orjson==3.9.10

# Math and scientific computing
numpy==1.24.3

# Optional: Database support if needed later
# sqlalchemy==2.0.23
# alembic==1.12.1

# Optional: Caching if needed later  
# redis==5.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and data processing
email-validator==2.1.0



